+++ examples/eventloop/c_eventloop.js
@@ -16,8 +16,14 @@ function setTimeout(func, delay) {
     var bind_args;
     var timer_id;
 
+    // Delay can be optional at least in some contexts, so tolerate that.
+    // https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/setTimeout
     if (typeof delay !== 'number') {
-        throw new TypeError('delay is not a number');
+        if (typeof delay === 'undefined') {
+            delay = 0;
+        } else {
+            throw new TypeError('invalid delay');
+        }
     }
 
     if (typeof func === 'string') {
@@ -53,7 +59,11 @@ function setInterval(func, delay) {
     var timer_id;
 
     if (typeof delay !== 'number') {
-        throw new TypeError('delay is not a number');
+        if (typeof delay === 'undefined') {
+            delay = 0;
+        } else {
+            throw new TypeError('invalid delay');
+        }
     }
 
     if (typeof func === 'string') {